Internal Interface

Documentation for LiveServer.jl's internal interface

File watching

There are two key types related to file watching:

one to wrap a file being watched (LiveServer.WatchedFile),

one for the file watcher itself wrapping the list of watched files and what to do upon file changes ("callback" function)

Any file watcher will be a subtype of the abstract type LiveServer.FileWatcher with, for instance, the default watcher being LiveServer.SimpleWatcher.

WatchedFile

LiveServer.WatchedFile — Type

WatchedFile

Struct for a file being watched containing the path to the file as well as the time of last modification.

LiveServer.has_changed — Function

has_changed(wf::WatchedFile)

Check if a WatchedFile has changed. Returns -1 if the file does not exist, 0 if it does exist but has not changed, and 1 if it has changed.

LiveServer.set_unchanged! — Function

set_unchanged!(wf::WatchedFile)

Set the current state of a WatchedFile as unchanged"

FileWatcher

Key types

LiveServer.FileWatcher — Type

FileWatcher

Abstract Type for file watching objects such as SimpleWatcher.

LiveServer.SimpleWatcher — Type

SimpleWatcher([callback]; sleeptime::Float64=0.1) <: FileWatcher

A simple file watcher. You can specify a callback function, receiving the path of each file that has changed as an AbstractString, at construction or later by the API function set_callback!. The sleeptime is the time waited between two runs of the loop looking for changed files, it is constrained to be at least 0.05s.

Functions related to a FileWatcher

LiveServer.start — Function

start(fw::FileWatcher)

Start the file watcher and wait to make sure the task has started.

LiveServer.stop — Function

stop(fw::FileWatcher)

Stop the file watcher. The list of files being watched is preserved and new files can still be added to the file watcher using watch_file!. It can be restarted with start. Returns a Bool indicating whether the watcher was running before stop was called.

LiveServer.set_callback! — Function

set_callback!(fw::FileWatcher, callback::Function)

Set or change the callback function being executed upon a file change. Can be "hot-swapped", i.e. while the file watcher is running.

LiveServer.watch_file! — Function

watch_file!(fw::FileWatcher, f_path::AbstractString)

Add a file to be watched for changes.

LiveServer.file_watcher_task! — Function

file_watcher_task!(w::FileWatcher)

Helper function that's spawned as an asynchronous task and checks for file changes. This task is normally terminated upon an InterruptException and shows a warning in the presence of any other exception.

Additional helper functions:

LiveServer.is_running — Function

is_running(fw::FileWatcher)

Checks whether the file watcher is running.

LiveServer.is_watched — Function

is_watched(fw::FileWatcher, f_path::AbstractString)

Checks whether the file specified by f_path is being watched.

Live serving

The serve method instantiates a listener (HTTP.listen) in an asynchronous task. The callback upon an incoming HTTP stream decides whether it is a standard HTTP request or a request for an upgrade to a websocket connection. The former case is handled by LiveServer.serve_file, the latter by LiveServer.ws_tracker. Finally, LiveServer.file_changed_callback is the function passed to the file watcher to be executed upon file changes.

LiveServer.serve_file — Function

serve_file(fw, req::HTTP.Request; inject_browser_reload_script::Bool = true)

Handler function for serving files. This takes a file watcher, to which files to be watched can be added, and a request (e.g. a path entered in a tab of the browser), and converts it to the appropriate file system path. If the path corresponds to a HTML file, it will inject the reloading <script> (see file client.html) at the end of its body, i.e. directly before the </body> tag. All files served are added to the file watcher, which is responsible to check whether they're already watched or not. Finally the file is served via a 200 (successful) response. If the file does not exist, a response with status 404 and an according message is sent.

LiveServer.ws_upgrade — Function

ws_upgrade(http::HTTP.Stream)

Upgrade the HTTP request in the stream to a websocket.

LiveServer.ws_tracker — Function

ws_tracker(ws::HTTP.WebSockets.WebSocket, target::AbstractString)

Adds the websocket connection to the viewers in the global dictionary WS_VIEWERS to the entry corresponding to the targeted file.

LiveServer.file_changed_callback — Function

file_changed_callback(f_path::AbstractString)

Function reacting to the change of the file at f_path. Is set as callback for the file watcher.

Additional helper functions:

LiveServer.get_fs_path — Function

get_fs_path(req_path::AbstractString)

Return the filesystem path corresponding to a requested path, or an empty String if the file was not found.

LiveServer.update_and_close_viewers! — Function

update_and_close_viewers!(wss::Vector{HTTP.WebSockets.WebSocket})

Take a list of viewers, i.e. WebSocket connections from a client, send a message with data "update" to each of them (to trigger a page reload), then close the connection. Finally, empty the list since all connections are closing anyway and clients will re-connect from the re-loaded page.

Helper functions associated with servedocs

LiveServer.servedocs_callback! — Function

servedocs_callback!(docwatcher, filepath, path2makejl, literate, foldername)

Custom callback used in servedocs triggered when the file corresponding to filepath is changed. If that file is docs/make.jl, the callback will check whether any new files have been added in the docs/src folder and add them to the watched files, it will also remove any file that may have been deleted or renamed. Otherwise, if the modified file is in docs/src or is docs/make.jl, a pass of Documenter is triggered to regenerate the documents, subsequently the LiveServer will render the produced pages in docs/build. foldername can be set other than "docs" if needed. buildfoldername can be set other than "build" if needed.

LiveServer.scan_docs! — Function

scan_docs!(dw::SimpleWatcher, literate="", foldername="docs")

Scans the docs/ folder in order to recover the path to all files that have to be watched and add those files to dw.watchedfiles. The function returns the path to docs/make.jl. A list of folders and file paths can also be given for files that should be watched in addition to the content of docs/src. foldername can be changed if it's different than docs.

Miscellaneous

LiveServer.example — Function

example()

Simple function to copy an example website folder to the current working directory that can be watched by the LiveServer to get an idea of how things work.

Example

LiveServer.example()
cd("example")
serve()

LiveServer.setverbose — Function

setverbose(b)

Set the verbosity of LiveServer to either true (showing messages upon events) or false (default).