Public Interface

Documentation for LiveServer.jl's exported functions

LiveServer.serve — Function

serve(filewatcher; host="127.0.0.1", port=8000, dir="", verbose=false, coreloopfun=(c,fw)->nothing, inject_browser_reload_script::Bool = true)

Main function to start a server at http://host:port and render what is in the current directory. (See also example for an example folder).

filewatcher is a file watcher implementing the API described for SimpleWatcher (which also is the default) and messaging the viewers (via WebSockets) upon detecting file changes.

port is an integer between 8000 (default) and 9000.

dir specifies where to launch the server if not the current working directory.

verbose is a boolean switch to make the server print information about file changes and connections.

coreloopfun specifies a function which can be run every 0.1 second while the liveserver is going; it takes two arguments: the cycle counter and the filewatcher. By default the coreloop does nothing.

Example

LiveServer.example()
serve(host="127.0.0.1", port=8080, dir="example", verbose=true)

If you open a browser to http://127.0.0.1:8080/, you should see the index.html page from the example folder being rendered. If you change the file, the browser will automatically reload the page and show the changes.

LiveServer.servedocs — Function

servedocs(; verbose=false, literate="", doc_env=false, foldername="docs")

Can be used when developing a package to run the docs/make.jl file from Documenter.jl and then serve the docs/build folder with LiveServer.jl. This function assumes you are in the directory [MyPackage].jl with a subfolder docs.

verbose=false is a boolean switch to make the server print information about file changes and

connections.

literate="" is the path to the folder containing the literate scripts, if left empty, it will be

assumed that they are in docs/src.

doc_env=false is a boolean switch to make the server start by activating the doc environment or not (i.e. the Project.toml in docs/).

skip_dir="" is a subpath of docs/ where modifications should not trigger the generation of the docs, this is useful for instance if you're using Weave and Weave generates some files in docs/src/examples in which case you should give skip_dir=joinpath("docs","src","examples").

skip_dirs=[] same as skip_dir but for a vector of such dirs. Takes precedence over skip_dir.

foldername="docs" specify the name of the content folder if different than "docs".

buildfoldername="build" specify the name of the build folder if different than "build".