Internals

Internal Documentation

Documentation for LiveServer.jl's internal interface

Internals for file watching

LiveServer.update_and_close_viewers! — Function.

update_and_close_viewers!(wss::Vector{HTTP.WebSockets.WebSocket})

Take a list of viewers, i.e. WebSocket connections from a client, send a message with data "update" to each of them (to trigger a page reload), then close the connection. Finally, empty the list since all connections are closing anyway and clients will re-connect from the re-loaded page.

source

Internals for live serving

LiveServer.ws_tracker — Function.

ws_tracker(::HTTP.Stream)

The websocket tracker. Upgrades the HTTP request in the stream to a websocket and adds this connection to the viewers in the global dictionary WS_VIEWERS.

source