Internals

Internal Interface

Documentation for LiveServer.jl's internal interface

File watching

There are two key types related to file watching:

one to wrap a file being watched (LiveServer.WatchedFile),

one for the file watcher itself wrapping the list of watched files and what to do upon file changes ("callback" function)

Any file watcher will be a subtype of the abstract type LiveServer.FileWatcher with, for instance, the default watcher being LiveServer.SimpleWatcher.

Watched file

LiveServer.WatchedFile — Type.

WatchedFile

Struct for a file being watched containing the path to the file as well as the time of last modification.

source

LiveServer.has_changed — Function.

has_changed(wf::WatchedFile)

Check if a WatchedFile has changed. Returns -1 if the file does not exist, 0 if it does exist but has not changed, and 1 if it has changed.

source

LiveServer.set_unchanged! — Function.

set_unchanged!(wf::WatchedFile)

Set the current state of a WatchedFile as unchanged"

source

File watcher

Key types

LiveServer.FileWatcher — Type.

FileWatcher

Abstract Type for file watching objects such as SimpleWatcher.

source

LiveServer.SimpleWatcher — Type.

SimpleWatcher([callback]; sleeptime::Float64=0.1) <: FileWatcher

A simple file watcher. You can specify a callback function, receiving the path of each file that has changed as an AbstractString, at construction or later by the API function described below. The sleeptime is the time waited between two runs of the loop looking for changed files, it is constrained to be at least 0.05s.

source

Functions related to a FileWatcher

LiveServer.start
LiveServer.stop
LiveServer.set_callback!
LiveServer.watch_file!
LiveServer.file_watcher_task!

Additional helper functions:

LiveServer.is_running — Function.

is_running(fw::FileWatcher)

Optional API function to check whether the file watcher is running.

source

LiveServer.is_watched — Function.

is_watched(fw::FileWatcher, f_path::AbstractString)

Check whether a file f_path is being watched by file watcher fw.

source

Live serving

The exported serve and verbose functions are not stated again. The serve method instantiates a listener (HTTP.listen) in an asynchronous task. The callback upon an incoming HTTP stream decides whether it is a standard HTTP request or a request for an upgrade to a websocket connection. The former case is handled by LiveServer.serve_file, the latter by LiveServer.ws_tracker. Finally, LiveServer.file_changed_callback is the function passed to the file watcher to be executed upon file changes.

LiveServer.serve_file — Function.

serve_file(fw, req::HTTP.Request)

Handler function for serving files. This takes a file watcher, to which files to be watched can be added, and a request (e.g. a path entered in a tab of the browser), and converts it to the appropriate file system path. If the path corresponds to a HTML file, it will inject the reloading script (see file client.js) at the end of its body, i.e. directly before the </body> tag. All files served are added to the file watcher, which is responsible to check whether they're already watched or not. Finally the file is served via a 200 (successful) response. If the file does not exist, a response with status 404 and message "404 not found" is sent.

source

LiveServer.ws_tracker — Function.

ws_tracker(ws::HTTP.WebSockets.WebSocket, target::AbstractString)

Adds the websocket connection to the viewers in the global dictionary WS_VIEWERS to the entry corresponding to the targeted file.

source

LiveServer.file_changed_callback — Function.

file_changed_callback(f_path::AbstractString)

Function reacting to the change of a file f_path. Is set as callback for the file watcher.

source

Additional helper functions:

LiveServer.get_fs_path — Function.

get_fs_path(req_path::AbstractString)

Return the filesystem path corresponding to a requested path, or an empty String if the file was not found.

source

LiveServer.update_and_close_viewers! — Function.

update_and_close_viewers!(wss::Vector{HTTP.WebSockets.WebSocket})

Take a list of viewers, i.e. WebSocket connections from a client, send a message with data "update" to each of them (to trigger a page reload), then close the connection. Finally, empty the list since all connections are closing anyway and clients will re-connect from the re-loaded page.

source

Miscellaneous

LiveServer.example — Function.

example()

Simple function to copy an example website folder to the current working directory that can be watched by the LiveServer to get an idea of how things work.

Example

LiveServer.example()
cd("example")
serve()

source