Public

Internal Documentation

Documentation for LiveServer.jl's exported functions

LiveServer.serve — Function.

serve(fw::FileWatcher=SimpleWatcher(); port::Int)

Main function to start a server at http://localhost:port and render what is in the current directory. (See also example for an example folder).

filewatcher is a file watcher implementing the API described for SimpleWatcher and

messaging the viewers (web sockets) upon detecting file changes.

port is an integer between 8000 (default) and 9000.

Example

LiveServer.example()
cd("example")
serve()

If you open a browser to http://localhost:8000, you should see the index.html page from the example folder being rendered. If you change the file, the browser will automatically reload the page and show the changes.

source

LiveServer.verbose — Function.

verbose(b)

Set the verbosity to either true (showing messages) or false (default).

source