Public

Public Interface

Documentation for LiveServer.jl's exported functions

LiveServer.serve — Function.

serve(filewatcher; port, directory)

Main function to start a server at http://localhost:port and render what is in the current directory. (See also example for an example folder).

filewatcher is a file watcher implementing the API described for SimpleWatcher and messaging the viewers (web sockets) upon detecting file changes.

port is an integer between 8000 (default) and 9000.

directory specifies where to launch the server if not the current working directory

Example

LiveServer.example()
serve(port=8080, dir="example")

If you open a browser to http://localhost:8080/, you should see the index.html page from the example folder being rendered. If you change the file, the browser will automatically reload the page and show the changes.

source

LiveServer.servedocs — Function.

servedocs()

Can be used when developping a package to run the docs/make.jl file from Documenter.jl and then serve the docs/build folder with LiveServer.jl. This function assumes you are in the directory [MyPackage].jl with a subfolder docs. Note: if you add new .md page, you will have to stop and restart servedocs.

source

LiveServer.verbose — Function.

verbose(b)

Set the verbosity of LiveServer to either true (showing messages upon events) or false (default).

source