Public

Public Interface

Documentation for LiveServer.jl's exported functions

LiveServer.serve — Function.

serve(filewatcher; port=8000, dir="", verbose=false, coreloop=(c,fw)->nothing)

Main function to start a server at http://localhost:port and render what is in the current directory. (See also example for an example folder).

filewatcher is a file watcher implementing the API described for SimpleWatcher (which also is the default) and messaging the viewers (via WebSockets) upon detecting file changes.

port is an integer between 8000 (default) and 9000.

dir specifies where to launch the server if not the current working directory.

verbose is a boolean switch to make the server print information about file changes and connections.

coreloopfun specifies a function which can be run every 0.1 second while the liveserver is going; it takes two arguments: the cycle counter and the filewatcher. By default the coreloop does nothing.

Example

LiveServer.example()
serve(port=8080, dir="example", verbose=true)

If you open a browser to http://localhost:8080/, you should see the index.html page from the example folder being rendered. If you change the file, the browser will automatically reload the page and show the changes.

source

LiveServer.servedocs — Function.

servedocs()

Can be used when developping a package to run the docs/make.jl file from Documenter.jl and then serve the docs/build folder with LiveServer.jl. This function assumes you are in the directory [MyPackage].jl with a subfolder docs.

source